chore(tests): strip redundant asyncio markers, add shared helpers, deselect integration by default
Remove 387 redundant @pytest.mark.asyncio decorators (asyncio_mode = "auto"
already handles them). Extract duplicated _FAKE_TOKEN/_make_credential/_make_client
helpers into tests/unit/services/_helpers.py and update 9 service test files to
import from there. Add -m 'not integration' to pyproject addopts so a bare pytest
skips integration; fix just integration to pass -m integration to re-select them.
Replace RuntimeError in integration/conftest.py workspace_id fixture with
pytest.skip(allow_module_level=True) for better local DX.
