+++ tests/unit/services/test_sql_exec.py
@@ -55,7 +55,6 @@ def _make_no_result_conn(*, rowcount: int = 1) -> MagicMock:
 # ---------------------------------------------------------------------------
 
 
-@pytest.mark.asyncio
 async def test_execute_select_returns_sql_result() -> None:
     target = _make_target()
     conn = _make_conn([(1, "hello"), (2, "world")], ["id", "name"])
@@ -66,7 +65,6 @@ async def test_execute_select_returns_sql_result() -> None:
     assert isinstance(result, SqlResult)
 
 
-@pytest.mark.asyncio
 async def test_execute_select_columns_and_rows() -> None:
     target = _make_target()
     conn = _make_conn([(42, "foo")], ["col_a", "col_b"])
@@ -78,7 +76,6 @@ async def test_execute_select_columns_and_rows() -> None:
     assert result.rows == [[42, "foo"]]
 
 
-@pytest.mark.asyncio
 async def test_execute_select_empty_rows() -> None:
     target = _make_target()
     conn = _make_conn([], ["col_a", "col_b"])
@@ -90,7 +87,6 @@ async def test_execute_select_empty_rows() -> None:
     assert result.rows == []
 
 
-@pytest.mark.asyncio
 async def test_execute_select_rowcount_falls_back_to_len_rows() -> None:
     """When driver returns rowcount=-1 for SELECT, we use len(rows)."""
     target = _make_target()
@@ -102,7 +98,6 @@ async def test_execute_select_rowcount_falls_back_to_len_rows() -> None:
     assert result.rowcount == 3
 
 
-@pytest.mark.asyncio
 async def test_execute_select_rowcount_from_driver_when_positive() -> None:
     """When driver returns a positive rowcount, use it directly."""
     target = _make_target()
@@ -119,7 +114,6 @@ async def test_execute_select_rowcount_from_driver_when_positive() -> None:
 # ---------------------------------------------------------------------------
 
 
-@pytest.mark.asyncio
 async def test_execute_insert_no_rows_returns_empty() -> None:
     target = _make_target()
     conn = _make_no_result_conn(rowcount=3)
@@ -132,7 +126,6 @@ async def test_execute_insert_no_rows_returns_empty() -> None:
     assert result.rowcount == 3
 
 
-@pytest.mark.asyncio
 async def test_execute_dml_closes_connection() -> None:
     target = _make_target()
     conn = _make_no_result_conn()
@@ -148,7 +141,6 @@ async def test_execute_dml_closes_connection() -> None:
 # ---------------------------------------------------------------------------
 
 
-@pytest.mark.asyncio
 async def test_execute_alter_no_rows_returns_empty() -> None:
     target = _make_target()
     conn = _make_no_result_conn(rowcount=0)
@@ -165,7 +157,6 @@ async def test_execute_alter_no_rows_returns_empty() -> None:
 # ---------------------------------------------------------------------------
 
 
-@pytest.mark.asyncio
 async def test_execute_datetime_column_serialised_to_iso() -> None:
     dt = datetime(2024, 6, 1, 12, 0, 0, tzinfo=UTC)
     target = _make_target()
@@ -177,7 +168,6 @@ async def test_execute_datetime_column_serialised_to_iso() -> None:
     assert result.rows[0][0] == dt.isoformat()
 
 
-@pytest.mark.asyncio
 async def test_execute_decimal_column_serialised_to_string() -> None:
     target = _make_target()
     conn = _make_conn([(Decimal("3.14"),)], ["price"])
@@ -188,7 +178,6 @@ async def test_execute_decimal_column_serialised_to_string() -> None:
     assert result.rows[0][0] == "3.14"
 
 
-@pytest.mark.asyncio
 async def test_execute_bytes_column_base64_encoded() -> None:
     raw = b"\x00\x01\x02\x03"
     target = _make_target()
@@ -201,7 +190,6 @@ async def test_execute_bytes_column_base64_encoded() -> None:
     assert result.rows[0][0] == expected_b64
 
 
-@pytest.mark.asyncio
 async def test_execute_bytes_column_name_gets_base64_suffix() -> None:
     target = _make_target()
     conn = _make_conn([(b"\xff",)], ["hash_val"])
@@ -212,7 +200,6 @@ async def test_execute_bytes_column_name_gets_base64_suffix() -> None:
     assert result.columns == ["hash_val__base64"]
 
 
-@pytest.mark.asyncio
 async def test_execute_non_binary_column_name_unchanged() -> None:
     target = _make_target()
     conn = _make_conn([(42,)], ["score"])
@@ -228,7 +215,6 @@ async def test_execute_non_binary_column_name_unchanged() -> None:
 # ---------------------------------------------------------------------------
 
 
-@pytest.mark.asyncio
 async def test_execute_syntax_error_propagates() -> None:
     """Non-mapped driver errors are raised as-is."""
     target = _make_target()
@@ -244,7 +230,6 @@ async def test_execute_syntax_error_propagates() -> None:
         await sql_exec.execute(target, "SLECT 1")
 
 
-@pytest.mark.asyncio
 async def test_execute_permission_denied_raises_permission_denied() -> None:
     target = _make_target()
     conn = MagicMock()
@@ -259,7 +244,6 @@ async def test_execute_permission_denied_raises_permission_denied() -> None:
         await sql_exec.execute(target, "SELECT * FROM SensitiveTable")
 
 
-@pytest.mark.asyncio
 async def test_execute_permission_denied_message_contains_hint() -> None:
     """PermissionDenied message must mention a documentation hint."""
     target = _make_target()
@@ -275,7 +259,6 @@ async def test_execute_permission_denied_message_contains_hint() -> None:
         await sql_exec.execute(target, "SELECT * FROM X")
 
 
-@pytest.mark.asyncio
 async def test_execute_auth_error_raises_auth_error() -> None:
     """Authentication failures (expired/missing token) are re-raised as AuthError."""
     target = _make_target()
@@ -291,7 +274,6 @@ async def test_execute_auth_error_raises_auth_error() -> None:
         await sql_exec.execute(target, "SELECT 1")
 
 
-@pytest.mark.asyncio
 async def test_execute_perm_denied_driver_raises_permission_denied() -> None:
     """SQL permission-denial errors are re-raised as PermissionDenied."""
     target = _make_target()
@@ -312,7 +294,6 @@ async def test_execute_perm_denied_driver_raises_permission_denied() -> None:
 # ---------------------------------------------------------------------------
 
 
-@pytest.mark.asyncio
 async def test_execute_multi_statement_returns_last_result_set() -> None:
     """When the cursor has multiple result sets, the last one is returned."""
     target = _make_target()
@@ -342,7 +323,6 @@ async def test_execute_multi_statement_returns_last_result_set() -> None:
 # ---------------------------------------------------------------------------
 
 
-@pytest.mark.asyncio
 async def test_execute_closes_connection_after_success() -> None:
     target = _make_target()
     conn = _make_conn([(1,)], ["n"])
@@ -353,7 +333,6 @@ async def test_execute_closes_connection_after_success() -> None:
     conn.close.assert_called_once()
 
 
-@pytest.mark.asyncio
 async def test_execute_closes_connection_on_error() -> None:
     """Connection must be closed even when cursor.execute raises."""
     target = _make_target()
@@ -376,7 +355,6 @@ async def test_execute_closes_connection_on_error() -> None:
 # ---------------------------------------------------------------------------
 
 
-@pytest.mark.asyncio
 async def test_binary_null_first_row_detected_via_later_rows() -> None:
     """If the first row has NULL for a binary column, later rows with bytes
     must still cause the column to be tagged.  The old first-row-only heuristic
@@ -391,7 +369,6 @@ async def test_binary_null_first_row_detected_via_later_rows() -> None:
     assert result.columns == ["blob__base64"]
 
 
-@pytest.mark.asyncio
 async def test_binary_all_null_rows_not_tagged() -> None:
     """A column that is NULL in every row is not binary and must not be tagged."""
     target = _make_target()
@@ -403,7 +380,6 @@ async def test_binary_all_null_rows_not_tagged() -> None:
     assert result.columns == ["blob"]
 
 
-@pytest.mark.asyncio
 async def test_binary_type_code_in_description_tags_column() -> None:
     """When cursor.description carries a type_code whose __name__ contains
     'binary', the column is tagged without scanning any row."""
